pyproject.toml
[project]
name = "univ-software-application"
version = "0.1.0"
description = "resume screening"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.2",
    "fastapi>=0.120.4",
    "google-genai>=0.3.0",
    "httpx>=0.28.1",
    "jina>=3.8.2",
    "pandas>=2.3.3",
    "pydantic-settings>=2.11.0",
    "pypdf>=6.1.3",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "torch>=2.9.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "ruff>=0.14.3",
]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
